Questions display properly as strings in ui
Handlebars are displaying the survey title, questions, and answers
properly on the ui.
However, when trying to 'answer' the questions, the data being sent back
to the server is not being passed in the appropriate format.
The id seems to be correct for the path, but the data needs modification